Fix Some Bugs and Add Some Functions

+++ src/main/java/com/hassan/pets/Contoller/ItemsController.java
@@ -38,7 +38,7 @@ public class ItemsController {
 
 
     @GetMapping("{itemId}")
-    public ItemRecord getById(@PathVariable int itemId) {
+    public ItemRecord getById(@PathVariable Long itemId) {
         return itemService.getById(itemId);
     }
 
@@ -50,7 +50,7 @@ public class ItemsController {
 
 
     @DeleteMapping("{itemId}")
-    public void deleteItem(@PathVariable int itemId){
+    public void deleteItem(@PathVariable Long itemId){
         itemService.deleteItem(itemId);
     }
