+++ src/main/java/com/hassan/pets/DTO/ItemRecord.java
@@ -2,6 +2,6 @@ package com.hassan.pets.DTO;
 
 import java.math.BigDecimal;
 
-public record ItemRecord( int itemId, String itemName, BigDecimal price, int stock,
+public record ItemRecord( Long itemId, String itemName, BigDecimal price, int stock,
                           byte[] itemImage, String categoryName, String categoryDescription) {
 }
