+++ src/main/java/com/hassan/pets/DTO/UserRecord.java
@@ -3,7 +3,7 @@ package com.hassan.pets.DTO;
 
 
 public record UserRecord
-        (int userId, String username,
+        (Long userId, String username,
          String password, String email,
          String address, String phoneNumber) {
 }
