+++ src/main/java/com/hassan/pets/repository/UserRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends JpaRepository<Users,Integer> {
+public interface UserRepo extends JpaRepository<Users,Long> {
 }
